批量删除----------查询数据同步-----------登录校验已完成

+++ src/main/java/com/hnjd/ssm/service/impl/IUserServiceImpl.java
@@ -2,11 +2,15 @@ package com.hnjd.ssm.service.impl;
 
 import com.hnjd.ssm.domain.User;
 import com.hnjd.ssm.mapper.UserMapper;
+import com.hnjd.ssm.query.PageResult;
+import com.hnjd.ssm.query.QueryObject;
+import com.hnjd.ssm.query.UserQueryObject;
 import com.hnjd.ssm.service.IUserService;
 import com.hnjd.ssm.util.UserContext;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.util.CollectionUtils;
 import org.springframework.util.ObjectUtils;
 import org.springframework.util.StringUtils;
 
@@ -14,6 +18,7 @@ import javax.servlet.ServletContext;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Paths;
+import java.util.Collections;
 import java.util.List;
 
 /**
@@ -38,18 +43,24 @@ public class IUserServiceImpl implements IUserService {
     }
 
     @Override
-    public int delete(Long id) {
+    public int delete(String[] ids) {
+        UserQueryObject uqo = new UserQueryObject();
+        uqo.setDeleteAll(ids);
         //删除用户前若该用户存在用户头像则先将用户头像图片删除
-        User user = userMapper.getById(id);
-        String headImg = user.getHeadImg();
-        if (!StringUtils.isEmpty(headImg) && !"defaultHeadImg.png".equals(user.getHeadImg())) {
-            try {
-                Files.deleteIfExists(Paths.get(servletContext.getRealPath("images/headImg/" + headImg)));
-            } catch (IOException e) {
-                e.printStackTrace();
+        List<User> users = this.query4List(uqo).getResult();
+        if (!CollectionUtils.isEmpty(users)) {
+            for (User user : users) {
+                String headImg = user.getHeadImg();
+                if (!StringUtils.isEmpty(headImg) && !"defaultHeadImg.png".equals(user.getHeadImg())) {
+                    try {
+                        Files.deleteIfExists(Paths.get(servletContext.getRealPath("images/headImg/" + headImg)));
+                    } catch (IOException e) {
+                        e.printStackTrace();
+                    }
+                }
             }
         }
-        return userMapper.deleteById(id);
+        return userMapper.deleteById(uqo);
     }
 
     @Override
@@ -60,24 +71,46 @@ public class IUserServiceImpl implements IUserService {
     @Transactional(readOnly = true)
     @Override
     public User get(Long id) {
-        return userMapper.getById(id);
+        if (!ObjectUtils.isEmpty(id)) {
+            UserQueryObject uqo = new UserQueryObject();
+            uqo.setId(id);
+            List<User> users = userMapper.queryForList(uqo);
+            if (!CollectionUtils.isEmpty(users)) {
+                return users.get(0);
+            }
+        }
+        return null;
     }
 
     @Transactional(readOnly = true)
     @Override
-    public List<User> listAll() {
+    public PageResult query4List(QueryObject userQueryObject) {
+        Integer rows = userMapper.queryForCount(userQueryObject);
+        if (rows == 0) {
+            return new PageResult(Collections.EMPTY_LIST, 0, 1, userQueryObject.getPageSize());
+        }
+        PageResult pageResult = new PageResult(null, rows, userQueryObject.getCurrentPage(), userQueryObject.getPageSize());
+        //保证当前页始终不能比总页数大
+        if (userQueryObject.getCurrentPage() > pageResult.getTotalPage()) {
+            userQueryObject.setCurrentPage(pageResult.getTotalPage());
+        }
+        List<User> users = userMapper.queryForList(userQueryObject);
+        pageResult.setResult(users);
+        return pageResult;
         //int a = 1/0;
-        return userMapper.selectAll();
     }
 
     @Transactional(readOnly = true)
     @Override
     public void login(String username, String password) {
-        User user = userMapper.login(username, password);
+        UserQueryObject uqo = new UserQueryObject();
+        uqo.setUsername(username);
+        uqo.setPassword(password);
+        List<User> users = userMapper.queryForList(uqo);
         //若登录后的对象为null则抛出异常
-        if (ObjectUtils.isEmpty(user)) {
+        if (CollectionUtils.isEmpty(users)) {
             throw new RuntimeException("账号密码错误!");
         }
-        UserContext.setCurrentUser(user);
+        UserContext.setCurrentUser(users.get(0));
     }
 }
